\name{ARIeeg-package}
\alias{ARIeeg-package}
\alias{ARIeeg}
\docType{package}
\title{
\packageTitle{ARIeeg}
}
\description{
\packageDescription{ARIeeg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ARIeeg}
\packageIndices{ARIeeg}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ARIeeg}

Maintainer: \packageMaintainer{ARIeeg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
