% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIeeg.R
\name{ARIeeg}
\alias{ARIeeg}
\title{Parametric ARI for EEG data}
\usage{
ARIeeg(data, alpha,alternative,timeS,dist,formula,variable, B, effect,...)
}
\arguments{
\item{data}{data}

\item{alpha}{alpha level}

\item{alternative}{alternative hypothesis}

\item{timeS}{time signal to select, default NULL, i.e. no selection}

\item{dist}{an double defining the maximal distance for adjacency of two channels}

\item{formula}{formula object defining the design of the model.}

\item{variable}{variables to use in the right part of the formula}

\item{B}{number of permutation, default 5000}

\item{eff}{effect of interest where apply ARI}
}
\value{
Returns a data.frame with number of true discoveries for each cluster
}
\description{
Performs ARI using permutation local test for EEG data
}
\author{
Angela Andreella
}
