% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIpermEEG.R
\name{ARIpermEEG}
\alias{ARIpermEEG}
\title{ARI Permutation-based for EEG data}
\usage{
ARIpermEEG(data, alpha, family, delta, alternative, timeS, dist, formula, variable, B, effect, model, ...)
}
\arguments{
\item{data}{data, default is NULL. You need to insert data or model}

\item{alpha}{alpha level}

\item{family}{which family for the confidence envelope? simes, finner, beta or higher.criticism. default is simes}

\item{delta}{do you want to consider at least delta size set?}

\item{alternative}{alternative hypothesis}

\item{timeS}{time signal to select, default NULL, i.e. no selection}

\item{dist}{an double defining the maximal distance for adjacency of two channels}

\item{formula}{formula object defining the design of the model.}

\item{variable}{variables to use in the right part of the formula}

\item{B}{number of permutation, default 5000}

\item{eff}{effect of interest where apply ARI}

\item{permuco4brain}{model, default is NULL}
}
\value{
Returns a data.frame with number of true discoveries for each cluster
}
\description{
Performs ARI using permutation local test for EEG data
}
\author{
Angela Andreella
}
