% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNullDistributionEEG.R
\name{plotNullDistributionEEG}
\alias{plotNullDistributionEEG}
\title{plot null distribution EEG}
\usage{
plotNullDistributionEEG(P,family,alpha, ct, path, name, delta)
}
\arguments{
\item{P}{pvalues matrix permutation}

\item{family}{which family for the confidence envelope? simes, finner, beta or higher.criticism. default is simes}

\item{alpha}{alpha level}

\item{ct}{set of thresholds}

\item{path}{path}

\item{name}{plot name}

\item{delta}{do you want to consider at least delta size set?}
}
\value{
Returns plot null distribution
}
\description{
plot null distribution EEG
}
\author{
Angela Andreella
}
