% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cluster_eeg.R
\name{summary_cluster_eeg}
\alias{summary_cluster_eeg}
\title{summary cluster EEG}
\usage{
summary_cluster_eeg(clusters,model,cv,ix,pvalues)
}
\arguments{
\item{clusters}{id cluster}

\item{model}{clusterMass model from permuco4brain package}

\item{cv}{critical vector}

\item{ix}{set of interest}

\item{pvalues}{pvalues matrix where rows indicate the permutations}

\item{eff}{effect of interest}
}
\value{
Returns a list with the following objects: discoveries number of discoveries in the set selected, cluster id, p-value
}
\description{
gives summary for each cluster EEG from clusterMass method
}
\author{
Angela Andreella
}
