% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_hommel_eeg.R
\name{summary_hommel_eeg}
\alias{summary_hommel_eeg}
\title{summary cluster EEG}
\usage{
summary_hommel_eeg(hommel,ix,alpha, clusters,eff)
}
\arguments{
\item{hommel}{hommel object from hommel package}

\item{ix}{set of interest}

\item{alpha}{alpha level}

\item{clusters}{id clusters}

\item{eff}{effect of interest}
}
\value{
Returns a list with the following objects: discoveries number of discoveries in the set selected, cluster id, p-value
}
\description{
gives summary for each cluster EEG from clusterMass method
}
\author{
Angela Andreella
}
