% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsTOlst.R
\name{utilsTOlst}
\alias{utilsTOlst}
\title{Create eeg_lst object from eegUtils object}
\usage{
utilsTOlst(data, reference, unit, resolution)
}
\arguments{
\item{data}{eegUtils data}

\item{reference}{reference}

\item{unit}{unit}

\item{resolution}{resolution}
}
\value{
Returns an eeg lst from eeguana package
}
\description{
uilds an eeg_lst object composed of two data.table`objects and one tibble. All three are linked by a unique identifier .id. Amplitude values and timestamps appear in the signal table. Triggers, blinks, artifact rejection markings, and other events logged by the EEG recording software appear in the events table. Segment information and recording IDs appear in the segments tibble.
}
\author{
Angela Andreella
}
